Changed uint16 to uint32 for consistency

+++ LambdaEngine/Include/Rendering/RenderGraph.h
@@ -89,7 +89,7 @@ namespace LambdaEngine
 				TextureView**	ppPerSubImageTextureViews;
 				Sampler**		ppSamplers;
 				uint32			TextureCount;
-				uint16			SamplerCount;
+				uint32			SamplerCount;
 				uint32			PerImageSubImageTextureViewCount;
 			} ExternalTextureUpdate;
